"number objects with methods"

+++ number_objects/number_objects/main.m
@@ -10,7 +10,7 @@ int main(int argc, const char * argv[])
         numFloat = [NSNumber numberWithFloat:147.665]; // assigning a value as an float to a object
         
         int x = [numInt integerValue]; //getting the integer value of the object
-        float y = [numFloat floatValue]; //getting the integer value of the object
+        float y = [numFloat floatValue]; //getting the float value of the object
         
         NSLog(@"%i and %f", x, y);
